Publish `@finos/perspective-esbuild-plugin`
