+++ examples/esbuild-example/README.md
@@ -1,0 +1,3 @@
+# `esbuild` Example
+
+Simple `esbuild` example using `perspective-viewer` and plugins.
