+++ examples/esbuild-example/package.json
@@ -1,0 +1,24 @@
+{
+    "name": "esbuild-example",
+    "private": true,
+    "version": "1.6.0",
+    "description": "An esbuild example app built using `@finos/perspective-viewer`.",
+    "scripts": {
+        "build": "node build.js",
+        "start": "yarn build && http-server dist"
+    },
+    "keywords": [],
+    "license": "Apache-2.0",
+    "dependencies": {
+        "@finos/perspective": "^1.6.0",
+        "@finos/perspective-viewer": "^1.6.0",
+        "@finos/perspective-viewer-d3fc": "^1.6.0",
+        "@finos/perspective-viewer-datagrid": "^1.6.0",
+        "@finos/perspective-viewer-openlayers": "^1.6.0"
+    },
+    "devDependencies": {
+        "@finos/perspective-esbuild-plugin": "^1.6.0",
+        "esbuild": "^0.13.12",
+        "http-server": "^0.11.1"
+    }
+}
