+++ examples/esbuild-example/src/index.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width,initial-scale=1">
+
+        <title>Perspective `esbuild` Example</title>
+
+        <link rel="preload" href="perspective.cpp.wasm" as="fetch" type="application/wasm" crossorigin="anonymous">
+        <link rel="preload" href="perspective_viewer_bg.wasm" as="fetch" type="application/wasm" crossorigin="anonymous">
+        <link rel="preload" href="superstore.arrow" as="fetch" type="arraybuffer" crossorigin="anonymous">
+        <link rel="preload" href="perspective.worker.js" as="fetch" type="application/javascript" crossorigin="anonymous">
+        <link rel="preload" href="editor.worker.js" as="fetch" type="application/javascript" crossorigin="anonymous">
+
+        <script type="module" src="index.js"></script>
+        <link rel="stylesheet" href="index.css" />
+    </head>
+    <body></body>
+</html>
