+++ examples/webpack-cross-origin/package.json
@@ -19,6 +19,12 @@
     },
     "devDependencies": {
         "@finos/perspective-webpack-plugin": "^1.6.0",
+        "css-loader": "^0.28.7",
+        "style-loader": "^0.18.2",
+        "webpack": "^5.14.0",
+        "webpack-cli": "^4.3.1",
+        "webpack-dev-server": "^3.11.2",
+        "html-webpack-plugin": "^5.1.0",
         "http-server": "^0.11.1"
     }
 }
