+++ packages/perspective-esbuild-plugin/build.js
