+++ packages/perspective-esbuild-plugin/external.js
