+++ packages/perspective-esbuild-plugin/ignore_fonts.js
