+++ packages/perspective-esbuild-plugin/wasm.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 
 exports.WasmPlugin = function WasmPlugin(inline) {
     function setup(build) {
-        build.onResolve({filter: /\.wasm$/}, (args) => {
+        build.onResolve({filter: /^\@finos\/perspective.+?\.wasm$/}, (args) => {
             if (
                 args.namespace === "wasm-stub" ||
                 args.namespace === "wasm-inline"
