+++ packages/perspective-jupyterlab/build.js
@@ -2,10 +2,10 @@ const cpy = require("cpy");
 
 const {lessLoader} = require("esbuild-plugin-less");
 
-const {WasmPlugin} = require("@finos/perspective-build/wasm");
-const {WorkerPlugin} = require("@finos/perspective-build/worker");
-const {UMDLoader} = require("@finos/perspective-build/umd");
-const {build} = require("@finos/perspective-build/build");
+const {WasmPlugin} = require("@finos/perspective-esbuild-plugin/wasm");
+const {WorkerPlugin} = require("@finos/perspective-esbuild-plugin/worker");
+const {UMDLoader} = require("@finos/perspective-esbuild-plugin/umd");
+const {build} = require("@finos/perspective-esbuild-plugin/build");
 
 const TEST_BUILD = {
     entryPoints: ["src/js/psp_widget.js"],
