+++ packages/perspective-jupyterlab/package.json
@@ -42,7 +42,7 @@
         "@lumino/widgets": "1.33.0"
     },
     "devDependencies": {
-        "@finos/perspective-build": "^1.6.0",
+        "@finos/perspective-esbuild-plugin": "^1.6.0",
         "@finos/perspective-test": "^1.6.0"
     },
     "jupyterlab": {
