+++ packages/perspective-viewer-datagrid/package.json
@@ -46,7 +46,7 @@
         "regular-table": "=0.5.6"
     },
     "devDependencies": {
-        "@finos/perspective-build": "^1.6.0",
+        "@finos/perspective-esbuild-plugin": "^1.6.0",
         "@finos/perspective-test": "^1.6.0"
     }
 }
