+++ packages/perspective/package.json
@@ -60,7 +60,7 @@
         "ws": "^6.1.2"
     },
     "devDependencies": {
-        "@finos/perspective-build": "^1.6.0",
+        "@finos/perspective-esbuild-plugin": "^1.6.0",
         "@finos/perspective-cpp": "^1.6.0",
         "jsverify": "^0.8.4",
         "lodash": "^4.17.4",
