+++ rust/perspective-viewer/package.json
@@ -58,6 +58,7 @@
         "monaco-editor-webpack-plugin": "3.1.0"
     },
     "devDependencies": {
-        "@finos/perspective-test": "^1.6.0"
+        "@finos/perspective-test": "^1.6.0",
+        "@finos/perspective-esbuild-plugin": "^1.6.0"
     }
 }
