+++ rust/perspective-viewer/src/rust/components/viewer.rs
@@ -6,6 +6,9 @@
 // of the Apache License 2.0.  The full license can be found in the LICENSE
 // file.
 
+// TODO yew `Properties`
+#![allow(dead_code)]
+
 use super::column_selector::ColumnSelector;
 use super::config_selector::ConfigSelector;
 use super::containers::split_panel::SplitPanel;
